// src/helpers/schema.tsx
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type IdType = "CC" | "CE" | "NIT";

export type Int8 = ColumnType<string, bigint | number | string, bigint | number | string>;

export type LocationType = "bodega" | "kiosk" | "tienda";

export type MoveType = "ADJUSTMENT" | "RECEIPT" | "RETURN" | "SALE" | "TRANSFER";

export type Numeric = ColumnType<string, number | string, number | string>;

export type PaymentMethod = "card" | "cash" | "credit" | "digital_wallet" | "hybrid" | "points";

export type PaymentTermsType = "CONTADO" | "CREDITO";

export type PointsTransactionType = "adjustment" | "bonus" | "earned" | "redeemed";

export type PurchaseOrderStatus = "cancelled" | "draft" | "received" | "sent";

export type Timestamp = ColumnType<Date, Date | string, Date | string>;
// ...
export interface Locations {
  address: string | null;
  code: string;
  createdAt: Generated<Timestamp>;
  description: string | null;
  id: Generated<number>;
  isActive: Generated<boolean>;
  locationType: Generated<LocationType>;
  name: string;
  updatedAt: Generated<Timestamp>;
}
// ...
export interface ProductLots {
  expiresOn: Timestamp | null;
  id: Generated<string>;
  lotCode: string | null;
  productUuid: string;
}

export interface ProductPrices {
  channel: string | null;
  createdAt: Generated<Timestamp | null>;
  effectiveFrom: Generated<Timestamp>;
  effectiveTo: Timestamp | null;
  id: Generated<number>;
  locationId: number | null;
  price: Numeric;
  productId: number;
}

export interface Products {
  barcode: string | null;
  category: string | null;
  categoryId: number | null;
  cost: Generated<Numeric | null>;
  createdAt: Generated<Timestamp | null>;
  daysofcovertarget: number | null;
  description: string | null;
  id: Generated<number>;
  imageurl: string | null;
  isActive: Generated<boolean>;
  maxqty: Numeric | null;
  minimumStock: Generated<number>;
  minqty: Numeric | null;
  name: string;
  preferredSupplierId: number | null;
  price: Numeric;
  promotionalLabel: string | null;
  reorderenabled: Generated<boolean | null>;
  sku: string | null;
  stockQuantity: Generated<number>;
  subcategoryId: number | null;
  supplierId: number | null;
  updatedAt: Generated<Timestamp | null>;
  uuid: string | null;
}
// ...
export interface StockMovements {
  createdAt: Generated<Timestamp>;
  id: Generated<string>;
  locationId: number | null;
  lotId: string;
  productId: number | null;
  productUuid: string;
  qty: Numeric;
  reason: string | null;
  ref: string | null;
  type: MoveType;
}
// ...
export const MoveTypeArrayValues: [MoveType, ...MoveType[]] = ["ADJUSTMENT","RECEIPT","RETURN","SALE","TRANSFER"];